Moved no unit insertion rewrite into NoAutoTupling
