+++ scalafix-core/shared/src/main/scala/scalafix/rule/ScalafixRules.scala
@@ -18,8 +18,7 @@ object ScalafixRules {
     Sbt1(index),
     ExplicitResultTypes(index),
     RemoveUnusedImports(index),
-    NoAutoTupling(index),
-    NoUnitInsertion(index)
+    NoAutoTupling(index)
   )
   def all(index: SemanticdbIndex): List[Rule] =
     syntax ++ semantic(index)
